User authentication

+++ public/javascripts/discussionBoard.js
@@ -15,21 +15,20 @@ $(document).ready(function() {
 
 $(function() {
 
-    $('#postMessage').on('click', '.submitPost', function(event) {
-        
+    $('#postMessage').submit(function(event) {
+
         event.preventDefault();
 
-        var inputName = $('#lonelinessPostName')
+        var inputName = $('#lonelinessPostName').attr('name')
         var inputBody = $('#lonelinessPostBody')
 
         $.ajax( {
             url: 'loneliness/post',
             method: 'POST',
             contentType: 'application/json',
-            data: JSON.stringify({ name: inputName.val(), textbody: inputBody.val() }),
+            data: JSON.stringify({ name: inputName, textbody: inputBody.val() }),
             success: function(message) {
                 console.log(message)
-                inputName.val('')
                 inputBody.val('')
                 var formattedDate = new Date(message.datetime_of_post);
                 var posts = $('#posts')
@@ -40,7 +39,7 @@ $(function() {
         })
     })
 
-    $('#postMessage').on('click', '.clearForm', function(event) {
+    $('#postMessage').on('click', '#clearForm', function(event) {
 
         event.preventDefault();
 
@@ -49,22 +48,64 @@ $(function() {
     })
 
     $('#posts').on('click', '.bubble-container', function() {
-        if (confirm("Do you want to delete this post?")) {
-            $.ajax({
-                url: 'loneliness/' + $(this).data('id'),
-                method: 'DELETE',
-                contentType: 'application/json',
-                success: function(response) {
-                    console.log(response)
-                }
-            })
-            $(this).remove();
-        }  
+        if (($('#lonelinessPostName').attr('name') === $(this).data('name')) || $('#lonelinessPostName').attr('name') === "Admin")
+        {
+            if (confirm("Do you want to delete this post?")) {
+                $.ajax({
+                    url: 'loneliness/' + $(this).data('id'),
+                    method: 'DELETE',
+                    contentType: 'application/json',
+                    success: function(response) {
+                        console.log(response)
+                    }
+                })
+                $(this).remove();
+            }
+        }
     });
+
+    $('#signUp').submit(function(event) {
+
+        event.preventDefault();
+
+        var username = $('#username').val()
+        var password = $('#password').val()
+        $('#username').val('');
+        $('#password').val('');
+
+        $.ajax( {
+            url: 'auth/login',
+            method: 'POST',
+            contentType: 'application/json',
+            data: JSON.stringify({ username: username, password: password }),
+            success: function(message) {
+                if (message === "success") {
+                    window.location.reload();
+                } else {
+                    alert("Sorry, you entered the wrong password! Please try again.")
+                }
+            }
+        })
+    });
+
+    $('#signUp').on('click', '#logoutUser', function(event) {
+
+        event.preventDefault();
+
+        $.ajax( {
+            url: 'auth/logout',
+            method: 'POST',
+            contentType: 'application/json',
+            success: function(message) {
+                window.location.reload();
+            }
+        })
+    })
+    
 })
 
 function createCard(id, username, body, formattedDate) {
-    var li = "<div class=\"bubble-container\" data-id=\""+ id +"\" style=\"width: 100%;\">" 
+    var li = "<div class=\"bubble-container\" data-name=\"" + username +"\" data-id=\""+ id +"\" style=\"width: 100%;\">" 
         + "<div class=\"bubble\">" 
         + "<h5 class=\"card-title small-caps\">" 
         + username
